refactored ui folders
